Remove repetition in base_io by using an Enum
The writer/reader classes were introducing huge amounts of repetition
into the codebase with methods like write_bool, write_byte, write_ubyte
being very similar, with only different struct type. This instead
introduces an enum class with all possible struct types and a single
method `write_value` or `read_value` taking in this enum value.

In similar spirit of repetition avoidance, this also removes methods
like write_varint, write_varshort, write_varlong (and read counterparts)
in favor of a single write_varint with max_bits attribute, or
write_varuint for unsigned values (and read counterparts).

With that, the enforce_range decorator was also removed since it wasn't
necessary anymore, it was actually hardly necessary in the original
implementation since struct will produce errors on invalid values for us
directly anyway, and the only reason for it's use before were slightly
better error messages, but all things considered it's not worth the
downsides.

This basically meant almost a full rewrite of the base_io file, and
things interacting with it. And while this is a little less convenient
solution for using the writer/reader classes, as instead of
`writer.read_bool()` we now need `writer.read_value(StructFromat.BOOL)`,
I think it's worth it for how much repetition was avoided.