Remove fundamentally flawed ReadInstruction class
The ReadInstruction class was supposed to be a way to handle hard-coding
dynamic reads (such as reading varints where we can't know how many
bytes will actually be read). This was done by having this class be a
container for sync,async pair of read functions which would be obtained
from the BaseSyncReader/BaseAsyncReader classes, and when they're
supposed to get executed, `self` will be bound to a specific concrete
class. This is supposed to be a solution for avoiding the need of having
2 almost exactly same methods on all ReadCapable classes, one being for
sync read and the other being for async read.

The issue with this approach is that using functions from the base
methods disregards any overridden functionality of the actual concrete
classes. There would technically be a viable way around this, which
works only because of how incredibly dynamic python is, being to store
the method names and use getattr on the reader when actually running
them.

This approach would be very hard to understand and unexpected, and it
would also cause issues when working with compounded read instructions
(instructions which depend on the result of some previous instruction
since it's passed as a parameter to this one). With full function
objects, this is doable with the use of partial, but even then it gets
very annoying and complicated with more complex logic. However if we
went with the method name approach, even just storing these arguments
would require another datastructure, with the need to somehow encode
which parameter should get obtained from previous instruction, making
the complexity of this system absolutely insane.

Compound read instructions were an issue even with functions, since
using them would often lead to having to define a full method for sync,
and another full method for async read, just to construct some more
complex read instruction, which completely defeats the purpose of these
read instructions, being to avoid repetition and the need of specifying
both async and sync versions of read for ReadCapable classes.