Move to using poetry dependency groups
This moves the current monolithic dependency structure whcih only
distinguishes between dev dependencies and main project dependencies to
a more modular structure.

The new dependency groups are: test,lint and release. By default, these
dependency groups will always get installed with `poetry install`,
however we override this for the release group, which should only
be installed manually, as most people probably won't need releasing
dependencies. If we do need them though, we can just use `poetry install
--with release` to include this optional group.

On top of this advantage, because our dependencies are now categoriezed
like this, if we don't need a specific group in a workflow, like in the
lint workflow where we don't need to be installing dependencies for unit
testing (pytest, coverage, ...), we can simply use `--without test`,
reducing the install time.

Unlike extras, these dependency groups are internal to just poetry, so
when releasing project, only the main group (main project's
dependencies) will be added, these groups therefore don't have any
effect on the end-user of the library.

Dependency groups are a relatively new feature in poetry, and were
introduced as a part of the 1.2 release. For more info about them, see
the documentation:
- https://python-poetry.org/docs/managing-dependencies/#dependency-groups
- https://python-poetry.org/docs/pyproject/#dependencies-and-dependency-groups
- https://python-poetry.org/docs/cli/#install