Remove fundamentally flawed rw_capable classes
The Read/Write capable classes were mainly here to handle packets for
us, by making each packet class be read-write capable, hence forcing it
to define methods that take in a base reader/writer and use these to
serialize/deserialize the given class.

The issue with this approach is the required code replication for
sync/async reading, as we can't always know how much bytes to read for
given packet (because of some dynamicness, such as varints, utf/ascii
strings, bytearrays, etc. where we read as much data as was specified).

Instead of this, the Minecraft packet format actually makes this easier
for us, because of the packet format it uses, which simply requires each
packet to first send over it's size, then the data (having the specified
size). This received data can then be received and converted into a
Buffer, which we can then be deserialized into a packet. Similarly all
packets will be serializable into a buffer, which we can then send as a
packet (prefixing it's size).

This handling can be implemented later on in the packet classes, but for
now, this commit just removes the current flawed approach.