Install test dependencies in validation workflow
When moving to poetry dependency groups, test dependencies were split to
their own group which was ignored for validation workflows as it didn't
seem necessary to install it. However because during validation we're
running pyright in that python environment, which also runs for our test
files, missing these test dependencies results in reports of invalid
imports as pyright can't find pytest.

We could theoretically bypass this, but it's not worth it as having
pytest actually installed also brings in it's stubs giving us proper
type hints when working with it.