Make connection classes into context managers
It makes sense to support using closable connections as context
managers, so that we can be sure we always close the used connection
after it was used.

This is very important as not closing connections will keep an open file
pointer which might eventually result in a very unexpected error of "Too
many open files" which is then very hard to track down.

Note that this change uses async context managers for the async
connections, even though they wouldn't technically be required as
there's nothing in the close methods that's async. However that might
change with a different implementation, and it would be hard to preserve
backwards compatibility afterwards, so it's a better idea to just have
that support already, in case we need it.

With the async connections using async context managers, we also convert
the close method to be async itself for the same reason (potential
usefulness in the future, if the implementation changes).