Consolidate logic for packet versioning into an ABC
This moves all of the (relatively complex) logic responsible for
obtaining the individual versioned packets from `mcproto/packets/map.py`
to `mcproto/utils/version_map.py` with the packet map simply inheriting
from this base class and specifying where the packets should be imported
from, what is considered as a packet, and identifying a unique composite
key for each packet class, being (PacketDirection, GameState,
PACKET_ID).

This kind of consolidation bundles away the cluttery logic behind
collecting all of the versioned implementations and leaves us with
a much simpler to understand core logic for the packet map.

However the primary reason for this consolidation wasn't just making
packet map file cleaner, but rather to create a reusable base class able
to handle any kind of versioned components, not just packets. This will
therefore be especially useful in the future, as data types will be
added (such as classes for NBT tags, entity metadata, UUIDs, etc.) which
may also change as the protocol specification changes, and so to make it
possible to support multiple versions, a similar version map will be
needed.