Increase the stack level for version fallback warnings
By default, warnings use a stack-level of 1, which means only the
context of the warning call would be shown (i.e. the function in which
this happened). This increases the stack-level to 3, as this is an
internal fuction, only expected to be called from other functions, which
are then potentially called by the user.