Remove package_dir option from towncrier config
The `package_dir` option specifies which folder the package (to be
imported by towncrier, for determining the name and the version) lives
in. However in our case, the package is in the root project directory,
named "mcproto", having this set to "mcproto", it would make towncrier
attempt to look in "mcproto/mcproto", which isn't correct.