Leave docs building to readthedocs
Don't specify full sphinx-build commands in readthedocs configuration,
let readthedocs handle that, only specify the preparation steps
(installing dependencies and poetry).

This is important because when readthedocs builds this page, some
environmental variables are set which furo theme picks up on, and for
example adds the version selector menu, or similar readthedocs features
to the built documentation, whcih wouldn't otherwise be there when
building ourselves.