Add base minecraft authentication system
This adds support for minecraft authentication utilizing the Mojang
authserver (API). This system does not yet include any way to actually
authenticate, only to manage the already authenticated account.

This system is technically a part of the "Yggdrasil" system, however the
regular Yggdrasil authentication (using plaintext username and password)
is not currently a part of this.

This commit also adds httpx as a dependency, as it's used to make the
necessary HTTP requests.