Add Yggdrasil authentication
This adds an implementation of the "Yggdrasil" system for
authentication, whcih uses plaintext username/email and password to
obtain the access token (Bearer), which can then be used to interract
with the mojang API.

The Yggdrasil system only supports unmigrated (non-Microsoft) accounts.
Migrated accounts will not be able to login with it.

WARNING: This system is entirely based on the wiki.vg documentation, and
it has NOT actually been tested. This is because I don't have access to
any non-migrated minecraft accounts, and so I don't have any viable
accounts to test this implementation with. However the wiki.vg
documentation is quite good and covers how the system work in great
detail, so this *should* work fine.