Configue pyright to assume python version 3.8
As we didn't previously specify which python version should pyright use,
it automatically defaulted to the latest one. This adds a setting to
make pyright assume we're usng python 3.8 (lowest supported python
version), allowing it to easily catch some mistakes, such as using a
fucntion that's not yet available in older python versions when
developing with a newer python version installed.

Doing this actually uncovered some mistakes, namely:
- `docs/conf.py` used `str.removesuffix` function was used, which is
  however a 3.9+ function.
- `docs/conf.py` was missing a `__future__.annotations` import, causing
  annotations like `list[str]` to fail on 3.8 (3.9+)
- `mcproto/types/chat.py` was importing `TypeAlias` from `typing`,
  however `TypeAlias` was only introduced into `typing` in 3.9, so to
  support 3.8, we should be importing from `typing_extensions`.
- `mcproto/types/chat.py` was using `list[RawChatMessageDict]` in a
  type-alias constant definition, which however wouldn't work on 3.8
  (list was made generic in 3.9). This was solved by making this
  annotation a string instead.