Make pyright report incompatible mehtod overrides (LSP breaking)
In OOP programming, we should generally try to follow "LSP": Liskov
Substitution Principle, which basically states that any class A,
inheriting from another class B should be able to be treated as that
class B, without issues.

So as an example, if our parent class (B) has a method `foo`, which
doesn't take any arguments, in the new class (A), we can override this
method, but we can't now make `foo` take an additional argument, as that
would mean this new class A can't be assumed to act like B, just with
extra features. Hence this would be violating LSP.

In the existing code-base, there were 4 violations like this found, 2 of
which were resolved, and 2 ignored:

- First 2 violations were found in the
  `tests/mcproto/protocol/helpers.py` file, the violation was about
  overriding Mock's `__call__`, which we don't care about, as in this
  case we break LSP intentionally, to make this class behave like the
  object we intent to mock.
- Second 2 violations were found in
  `tests/mcproto/protocol/test_base_io.py` file, which were about our
  "concrete" writer classes, that directly inherit from the base writer
  classes, and override the abstract `write` function to make these
  classes initializable and testable. However this override actually
  mistakenly assumed that the `write` function was taking in a
  `bytearray`, when in fact the base class specifies that it takes
  `bytes`. This violation was fixed.

Additionally, during fixing of the second 2 violations, I have noticed
that in the `tests/mcproto/protocol/helpers.py` file (where the first
2 violations were), in the `WriteFunctionMock` class, the same mistake
was made (having the write function take `bytearray` instead of
`bytes`), which I fixed here as well.