Move the Installation page from README to docs installation page
Note that this also removes the git module installation option
instructions, as this option is not somemthing most people will utilize.
Git modules don't make much sense with python packages, which have their
own dependency management tool (pip). The only advantage with git
modules is that it's easier to understand where the source code is, and
to see it, however this isn't really too important for us, and it's
quite unnecessary.