Enable PL linter rules
This enables all ruff's pylint (PL) rules.

However PLR2004 (using unnamed numerical constants) is ignored, as these
kinds of constants are used all over the codebase, in comparisons, etc.
As mcproto is a fairly low level library, this is not surprising, and in
most cases, it's easy to understand what the numerical constant is in
the context of that comparison.