Enable PGH linter rules
This enables all ruff's pygrep-hooks (PGH) rules.

However PGH003 (using specific rule codes in type ignores) is ignored,
as we're using pyright rather than mypy, and while there are some
rulecodes for pyright and we could be more specific, most of the
violations just fall under the same (general typing issues) category,
which isn't very helpful for categorizing the issue. Because of that,
this would just be annoying to do without any huge benefit for it.