Fix packet compression handling
This fixes a bug in how we previously handled packet compression.
Originally, our handling was done based on a boolean, either do enable
compression, or don't.

However this doesn't really work, as the way we recognize whether to
enable compression is through the server, sending us the
`LoginSetCompression` packet, which only contains a compression
threshold value. This value can be -1, in which case compression should
be disabled, it can also be 0, in which case compression would indeed
always be enabled, but for any other value, compression should only be
enabled if the packet length surpasses this threshold, and we only know
the packet length during writing, so our interaction functions need to
be aware of it.

Note that if this threshold isn't crossed, but is set, the packet format
still changes (there is an extra varint after packet length, which
generally contains the uncompressed length, but when compression wasn't
used - below threshold, it will be set to 0, and the packet id and data
fields below remain uncompressed).

This changes `sync_write_packet` and `async_write_packet` functions and
replaces compressed bool parameter to compression_threshold parameter,
which is now being used properly. This however breaks compatibility, as
these writer functions are a part of the public API, and the compressed
parameter will no longer be available. Considering it didn't really even
work properly before, it's unlikely that someone was actually using it
yet, nevertheless it is a breaking change.

This does also affect how packet reading works, however there is no
breaking change, as only the logic needed to change, to account for the
possibility of the data length being 0, in which case compression wasn't
used. This function still only takes a compressed bool flag parameter,
rather than a compression_threshold, as if the threshold is
non-negative, the packet format fundamentally changes, so we should
already be reading the packets differently. For that reason, a bool flag
is sufficient here, and should just be set to True whenever the
threshold is >= 0.