Switch to 'all' for type-checking mode
This also modifies the diagnositic behavior for various circumstances.

This currently leaves all of the rules that would cause errors
explicitly disabled, while leaving everything else enabled.

Over time, the rules which were disabled for this reason will be
re-enabled and the corresponding issues will be fixed in future
commits.