Fix deprecation logic (#330)
Currently, the logic for handling deprecation is flawed, as it's using
`__package__` to obtain the currently installed version of `mcproto`,
however, since the deprecation.py file exists in a subpackage
(`mcproto.utils`), using `importlib.metadata.version` always fails with
`PackageNotFound`.

This was not caught in the unit-tests, because we blindly monkey-patched
the entire `importlib.metadata.version` function to always return the
expected version, regardless of the name of the package being checked.

This bug meant that the logic always assumed a fallback version of
v0.0.0, which happened silently and could've meant that using a
deprecated feature past it's specified deprecation version would still
only result in a deprecation warning, as opposed to a full runtime
exception getting raised.

This fixes the behavior by hard-coding the package-name to `mcproto` and
the monkey-patching in tests now only applies to `mcproto` package
specifically. Additionally, if we do get a `PackageNotFound` for
`mcproto` now, instead of silently ignoring it by assuming version
v0.0.0, a `RuntimeWarning` will be produced informing the user that such
a fallback has occurred.