Use codeclimate reporter id from gh secrets
Even though codeclimate doesn't consider the reporter id to be secret,
as it doesn't allow accessing any sensitive information and it's
write-only, it's still possible to use this id to maliciously upload
coverage from other projects as coverage for this project.

There isn't much of a reason why anyone would want to do this, all it
would achieve is potentially messing up the project status page that
shows the coverage score. But just the fact that this is possible is
still uncomfortable in my opinion.

For that reason, the coverage id for this project has been regenerated
and moved to github secrets.