Only run docs preview deploy on native branches/push
The original workflow somewhat naively ran deploy and always tried to
push the built docs to the repo pages, which is however very unsafe to
do from forks, not to mention that it wouldn't work anyways, since write
access & secrets aren't available when the workflow is ran from forks on
the `pull_request` trigger.

It is a little annoying that we can't safely show the docs preview page
for fork PRs, but it's the only safe way to go about this, and the docs
building will still occur, to make sure that the building doesn't fail.