Publish to Test PyPI with every commit to main
The `publish` workflow was updated to trigger package building and
publishing to Test PyPI with every commit to `main`.

This will provide us with a nice way to test out this CI workflow and
our build chain, giving us more confidence that when we'll want to do an
actual release, everything will work as it should.